Convert Menu to SCSS styling
