+++ src/components/menu/Menu/index.module.scss
@@ -49,8 +49,9 @@
   top: 8rem;
   right: 5rem;
   position: absolute;
+  width: 15rem;
 
-  width: 15%;
+  text-align: center;
   padding: 1em;
   background: $color-primary-darker;
 
@@ -61,14 +62,23 @@
 
   box-shadow: $box-shadow;
   border-radius: 10px;
+}
 
-  @include down($breakpoint-md) {
-    width: 25%;
-  }
+.menuBurgerContainer {
+  position: absolute;
+  top: 0;
+  left: 0;
+}
 
-  @include down($breakpoint-sm) {
-    width: 50%;
-  }
+.link {
+  color: $text-color-contrast;
+  width: 100%;
+}
+
+.link:hover {
+  text-decoration: none;
+  filter: brightness(85%);
+  cursor: pointer;
 }
 
 .menuLine {
@@ -77,8 +87,7 @@
   height: 7px;
 }
 
-.menuBurgerContainer {
-  position: absolute;
-  top: 0;
-  left: 0;
+.menuItem {
+  padding: 2em;
+  text-align: center;
 }
